only displaying column labels for now.
we can display both row and column labels on the image provided each row is a design element and we don't want to hyperlink it.  we are actually hyperlinking genes, but you won't see the gene names for some data sets like GSE2, because it does not contain any gene data.