Implements caching of an ontology to disk
If an ontology fails to load from its URL it will now attempt to load a
cached version from disk.

Successfully loading an ontology from URL will save a Jena processed
version of the RDF/XML file to disk as a cache in a folder specified
through the configuration ontology.cache.dir.