Fix ppavlidis/Gemma#2 - automatic reindexing of ontologies on load
Alters how ontologies are cached to disk. Previously we were caching a
Jena processed RDF/XML, now we are caching the original download.

The reason for this is related to the means of automatic reindexing. We
check each ontology on load to see if the newly created disk cache
differs from the previously created one. The Jena processed RDF/XML does
not guarantee any sort of ordering and thus would require a more complex
means of comparison than straight byte-by-byte.