Fix usage of "has part" and "has proper part" to determine children
List all the subjects X of the statements of the form where Y is the
class from which children are retrieved:

    X subClassOf <partOf|properPartOf Y>

Use OntModel methods instead of needlessly caching all the resources in the ontology service.

Always use transitive inference and remove the OntModelSpec argument
from the OntologyLoader API.

Add Uberon-based test cases.