Add efficient method for gathering children and parents of multiple terms
This is especially true for children as we only need to walk once
through the statement to gather those via "part of" and "proper part of".

Parents no longer include owl:Thing and children owl:Nothing nor
deprecated terms. There's a new option to retain them.

Return an ExtendedIterator in OntologySearch so that the service can
perform conversion and filtering of obsolete terms itself.

Move Jena-specific implementation details under ubic.basecode.ontology.jena

This package is now considered implementation details and should not be
used externally.